Creating cloud storage service for NCI. Requires some refactoring of current CloudStorageService First steps, does not compile yet.
