+++ src/main/java/org/auscope/portal/core/services/cloud/CloudComputeService.java
@@ -19,6 +19,8 @@ abstract public class CloudComputeService {
         NovaKeystone,
         /** Connect to an Amazon Web Services instance via EC2 */
         AWSEc2,
+        /** Connect to Raijin NCI HPC */
+        RAIJIN
     }
 
     /**
